# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestão Estoque CDD Maceió
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string(),
  }),
});

/**
 * @summary Admin logout
 */
export const AdminLogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current admin session
 */
export const GetAuthMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
});

/**
 * @summary Get stock items (public)
 */
export const getEstoqueQueryPageDefault = 1;
export const getEstoqueQueryLimitDefault = 20;

export const GetEstoqueQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  marca: zod.coerce.string().optional(),
  status: zod.enum(["OK", "NOK"]).optional(),
  curva: zod.enum(["A", "B", "C"]).optional(),
  page: zod.coerce.number().default(getEstoqueQueryPageDefault),
  limit: zod.coerce.number().default(getEstoqueQueryLimitDefault),
});

export const GetEstoqueResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      produto: zod.string(),
      marca: zod.string(),
      embalagem: zod.string(),
      doi: zod.number(),
      status: zod.enum(["OK", "NOK"]),
      demanda: zod.number(),
      min: zod.number(),
      max: zod.number(),
      curva: zod.enum(["A", "B", "C"]),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Get dashboard KPIs and charts (admin)
 */
export const GetEstoqueDashboardResponse = zod.object({
  totalSKUs: zod.number(),
  percentOK: zod.number(),
  percentNOK: zod.number(),
  mediaDOI: zod.number(),
  distribuicaoStatus: zod.array(
    zod.object({
      name: zod.string(),
      value: zod.number(),
    }),
  ),
  produtosPorMarca: zod.array(
    zod.object({
      marca: zod.string(),
      total: zod.number(),
      curva: zod.string(),
    }),
  ),
  snapshotDate: zod.string().nullable(),
});

/**
 * @summary Get list of all unique brands in current snapshot
 */
export const GetEstoqueMarcasResponse = zod.object({
  marcas: zod.array(zod.string()),
});

/**
 * @summary Get last upload status (admin)
 */
export const GetEstoqueUploadStatusResponse = zod.object({
  fileName: zod.string().nullable(),
  uploadedAt: zod.string().nullable(),
  totalRows: zod.number().nullable(),
});

/**
 * @summary Upload stock spreadsheet (admin)
 */
export const UploadEstoqueBody = zod.object({
  fileName: zod.string(),
  fileBase64: zod.string().describe("Base64 encoded file content"),
});

export const UploadEstoqueResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  rowsProcessed: zod.number(),
  fileName: zod.string(),
});