# lib/api-zod/src/generated/types/estoqueListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestão Estoque CDD Maceió
 * OpenAPI spec version: 0.1.0
 */
import type { EstoqueItem } from "./estoqueItem";

export interface EstoqueListResponse {
  items: EstoqueItem[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}