# lib/api-zod/src/generated/types/uploadEstoqueBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestão Estoque CDD Maceió
 * OpenAPI spec version: 0.1.0
 */

export interface UploadEstoqueBody {
  fileName: string;
  /** Base64 encoded file content */
  fileBase64: string;
}