# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestão Estoque CDD Maceió
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  success: boolean;
  message: string;
  rowsProcessed: number;
  fileName: string;
}