# lib/api-zod/src/generated/types/uploadStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestão Estoque CDD Maceió
 * OpenAPI spec version: 0.1.0
 */

export interface UploadStatus {
  fileName: string | null;
  uploadedAt: string | null;
  totalRows: number | null;
}